requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
duckdb==1.1.0
networkx==3.3
sentence-transformers==3.0.1
langdetect==1.0.9
groq==0.9.0
numpy<2.0.0
python-multipart==0.0.9
python-dotenv==1.0.1

bertopic==0.16.4
umap-learn==0.5.6
scikit-learn==1.5.1

datamapplot
matplotlib==3.9.2
